Moving classes to more logical packages
