+++ src/com/morgan/design/activity/fragment/BrewStatsFragments.java
@@ -1,4 +1,4 @@
-package com.morgan.design.fragment.refactor;
+package com.morgan.design.activity.fragment;
 
 import java.util.List;
